backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
razorpay==2.0.0
reportlab==4.4.10
python-dotenv==1.2.2
dj-database-url==2.3.0
whitenoise==6.12.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
requests==2.32.5
Pillow==12.2.0
